refactor(modpack): split import/export helpers and isolate format-specific exporters
