+++ src-tauri/src/instance/helpers/modpack/mod.rs
@@ -1,5 +1,5 @@
 pub mod curseforge;
 pub mod export;
-pub mod misc;
+pub mod import;
 pub mod modrinth;
 pub mod multimc;
