+++ src-tauri/src/instance/helpers/modpack/multimc.rs
@@ -1,6 +1,9 @@
 use crate::error::SJMCLResult;
-use crate::instance::helpers::modpack::misc::{ModpackManifest, ModpackMetaInfo};
-use crate::instance::models::misc::{InstanceError, ModLoader, ModLoaderType};
+use crate::instance::helpers::modpack::export::{
+  normalize_mod_loader_version, ExportModpackOptions, ModpackExportBundle,
+};
+use crate::instance::helpers::modpack::import::{ModpackManifest, ModpackMetaInfo};
+use crate::instance::models::misc::{Instance, InstanceError, ModLoader, ModLoaderType};
 use crate::resource::models::OtherResourceSource;
 use crate::tasks::PTaskParam;
 use async_trait::async_trait;
@@ -9,7 +12,7 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::Read;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 use tauri::AppHandle;
 use zip::ZipArchive;
 
@@ -169,4 +172,84 @@ fn get_version(component: &MultiMcComponent) -> SJMCLResult<String> {
     .or(component.cached_version.as_ref())
     .cloned()
     .ok_or(InstanceError::ModpackManifestParseError.into())
+}
+
+pub fn build_multimc_export_bundle(
+  instance: &Instance,
+  options: &ExportModpackOptions,
+  selected_files: &[(String, PathBuf)],
+) -> SJMCLResult<ModpackExportBundle> {
+  let manifest = generate_multimc_manifest(instance);
+  let json = serde_json::to_string_pretty(&manifest)
+    .map_err(|_| InstanceError::ModpackManifestParseError)?;
+
+  Ok(ModpackExportBundle {
+    overrides_prefix: ".minecraft".to_string(),
+    overrides_files: selected_files.to_vec(),
+    extra_files: vec![
+      ("mmc-pack.json".to_string(), json),
+      (
+        "instance.cfg".to_string(),
+        generate_multimc_instance_cfg(options),
+      ),
+      (".packignore".to_string(), String::new()),
+    ],
+  })
+}
+
+fn generate_multimc_manifest(instance: &Instance) -> MultiMcManifest {
+  let mut components = Vec::new();
+  components.push(MultiMcComponent {
+    uid: "net.minecraft".to_string(),
+    version: Some(instance.version.clone()),
+    important: Some(true),
+    dependency_only: Some(false),
+    cached_name: None,
+    cached_requires: None,
+    cached_version: None,
+    cached_volatile: None,
+  });
+
+  if instance.mod_loader.loader_type != ModLoaderType::Unknown {
+    let version = normalize_mod_loader_version(&instance.mod_loader.version);
+    let uid = match instance.mod_loader.loader_type {
+      ModLoaderType::Forge | ModLoaderType::LegacyForge => "net.minecraftforge",
+      ModLoaderType::NeoForge => "net.neoforged",
+      ModLoaderType::Fabric => "net.fabricmc.fabric-loader",
+      ModLoaderType::Quilt => "org.quiltmc.quilt-loader",
+      _ => "",
+    };
+
+    if !uid.is_empty() {
+      components.push(MultiMcComponent {
+        uid: uid.to_string(),
+        version: Some(version),
+        important: Some(false),
+        dependency_only: Some(false),
+        cached_name: None,
+        cached_requires: None,
+        cached_version: None,
+        cached_volatile: None,
+      });
+    }
+  }
+
+  MultiMcManifest {
+    format_version: 1,
+    components,
+    cfg: HashMap::new(),
+    base_path: String::new(),
+  }
+}
+
+fn generate_multimc_instance_cfg(options: &ExportModpackOptions) -> String {
+  let mut content = format!(
+    "# Auto generated by SJMC Launcher\nInstanceType=OneSix\nname={}\n",
+    options.name
+  );
+  if let Some(min_memory) = options.min_memory {
+    content.push_str("OverrideMemory=true\n");
+    content.push_str(&format!("MinMemAlloc={}\n", min_memory));
+  }
+  content
 }
